pyproject.toml
[project]
name = "tech-fault-rag-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.1",
    "openai>=2.29.0",
    "pypdf>=6.9.1",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "uvicorn>=0.42.0",
]

[tool.uv.workspace]
members = [
    "tech-fault-rag-bot",
]
